feat: 랜딩 페이지 새 페이지로 교체
