+++ src/pages/index.page.tsx
@@ -1,30 +1,65 @@
+'use client';
+
 import { useEffect, useState } from 'react';
+import { type GetServerSideProps } from 'next';
 import { useRouter } from 'next/router';
-import Image from 'next/image';
 import { useTranslation } from 'next-i18next';
+import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { useSetRecoilState } from 'recoil';
+import Spline from '@splinetool/react-spline';
+import { Application } from '@splinetool/runtime';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faPalette, faShareNodes } from '@fortawesome/free-solid-svg-icons';
+import {
+  faChevronDown,
+  faShareNodes,
+  faSwatchbook,
+} from '@fortawesome/free-solid-svg-icons';
+import { useIntersectionObserver } from '@Base/hooks/useIntersectionObserver';
 import { useCountUp } from '@Base/hooks/useCountUp';
 import { CropImage } from '@Recoil/app';
-import ColorChipSpinner from '@Components/ColorChipSpinner';
 import omctDb from '@Utils/omctDb';
 import { canWebShare, webShare } from '@Utils/share';
-import ROUTE_PATH from '@Constant/routePath';
 import { copyUrl } from '@Utils/clipboard';
-import questionBubble from 'public/images/icon/question-bubble.png';
+import ROUTE_PATH from '@Constant/routePath';
 import * as S from './style';
 
-import { GetServerSideProps } from 'next';
-import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
-
-function LandingPage() {
-  const { t, i18n } = useTranslation('common');
-
+export default function Home() {
   const [numberOfUsers, setNumberOfUsers] = useState(0);
+  const [isInViewBottom, setIsInViewBottom] = useState(false);
+  const [isRunningStartTransition, setIsRunningStartTransition] =
+    useState(false);
   const router = useRouter();
 
   const setUserImg = useSetRecoilState(CropImage);
+
+  const { t, i18n } = useTranslation('common');
+  const currentLocale = i18n.language;
+
+  const { ref: topRef } = useIntersectionObserver(
+    (entry: IntersectionObserverEntry) => {
+      if (entry.isIntersecting) setIsInViewBottom(false);
+    }
+  );
+  const { ref: bottomRef } = useIntersectionObserver(
+    (entry: IntersectionObserverEntry) => {
+      if (entry.isIntersecting) setIsInViewBottom(entry.isIntersecting);
+    }
+  );
+
+  const COUNT_UP_DURATION = 2000;
+  const userCount = useCountUp(
+    numberOfUsers,
+    COUNT_UP_DURATION,
+    isInViewBottom
+  );
+
+  useEffect(() => {
+    if (window !== undefined) document.body.style.backgroundColor = S.bgColor;
+
+    return () => {
+      if (window !== undefined) document.body.style.backgroundColor = '';
+    };
+  }, []);
 
   useEffect(() => {
     const getNumberOfUsers = async () => {
@@ -34,15 +69,24 @@ function LandingPage() {
     getNumberOfUsers();
   }, []);
 
-  const COUNT_UP_DURATION = 2000;
-  const count = useCountUp(numberOfUsers, COUNT_UP_DURATION, true);
-
   useEffect(() => {
     setUserImg('');
   }, [setUserImg]);
 
-  const onClickStartButton = () => {
+  const handleClickStart = () => {
+    setIsRunningStartTransition(true);
+  };
+
+  const onStartTransitionEnd = () => {
     router.push(ROUTE_PATH.imageUpload);
+  };
+
+  const handleScroll = () => {
+    if (typeof window !== 'undefined')
+      scrollTo({
+        top: isInViewBottom ? 0 : document.body.scrollHeight,
+        behavior: 'smooth',
+      });
   };
 
   const handleViewAllType = () => {
@@ -55,62 +99,76 @@ function LandingPage() {
     alert(t(`${messageId}`));
   };
 
-  const currentLocale = i18n.language;
-
-  const toggleLocale = currentLocale === 'en' ? 'ko-KR' : 'en';
-
   const handleToggleLanguage = () => {
-    router.push(router.pathname, router.asPath, { locale: toggleLocale });
+    router.push(router.pathname, router.asPath, {
+      locale: currentLocale === 'en' ? 'ko-KR' : 'en',
+    });
+  };
+
+  const onLoadSpline = (app: Application) => {
+    app.setZoom(0.15);
+    app.setBackgroundColor(S.bgColor);
   };
 
   return (
-    <>
-      <S.LandingWrap>
-        <S.LandingTitleDiv>
-          <S.LandingTitle>
-            {t('landingTitle_1')}{' '}
-            <S.TitleHighlight>{t('titleHighlight')}</S.TitleHighlight>{' '}
-            {t('landingTitle_2')}
-          </S.LandingTitle>
-          <S.LandingSubTitle>{t('landingSubTitle')}</S.LandingSubTitle>
-        </S.LandingTitleDiv>
+    <S.PageWrapper>
+      <S.TopPage>
+        <S.LanguageButton onClick={handleToggleLanguage}>
+          {{ 'ko-KR': 'ENG', 'en-US': '한국어' }[currentLocale]}
+        </S.LanguageButton>
 
-        <S.SpinnerWrapper>
-          <ColorChipSpinner />
-          <S.QuestionMark>
-            <Image
-              src={questionBubble}
-              alt="thought bubble"
-              width={48}
-              height={48}
-            />
-          </S.QuestionMark>
-        </S.SpinnerWrapper>
+        <S.TitleWrapper ref={topRef}>
+          <S.Title>{t('landingTitle')}</S.Title>
+          <S.Subtitle>{t('landingSubTitle')}</S.Subtitle>
+        </S.TitleWrapper>
 
-        <S.LandingBottomDiv>
-          <S.UserCount>
-            {t('userCount_1')} {count.toLocaleString()} {t('userCount_2')}
-            {t('userCount_3')}
-          </S.UserCount>
+        <Spline
+          scene="https://prod.spline.design/hZ8jGzxERn1B-4De/scene.splinecode"
+          style={{
+            width: '100%',
+            height: '100%',
+          }}
+          onLoad={onLoadSpline}
+        />
 
-          <S.StartButton onClick={onClickStartButton}>
-            {t('startButton')}
-          </S.StartButton>
+        <S.ScrollButton $isInViewBottom={isInViewBottom} onClick={handleScroll}>
+          <FontAwesomeIcon icon={faChevronDown} />
+        </S.ScrollButton>
+      </S.TopPage>
 
-          <S.MiniButtonWrapper>
-            <S.MiniButton onClick={handleViewAllType}>
-              <FontAwesomeIcon icon={faPalette} />
-            </S.MiniButton>
-            <S.MiniButton onClick={handleToggleLanguage}>
-              {{ 'ko-KR': 'ENG', en: '한국어' }[currentLocale]}
-            </S.MiniButton>
-            <S.MiniButton onClick={handleShare}>
-              <FontAwesomeIcon icon={faShareNodes} />
-            </S.MiniButton>
-          </S.MiniButtonWrapper>
-        </S.LandingBottomDiv>
-      </S.LandingWrap>
-    </>
+      <S.BottomPage>
+        <S.UserCountMessage ref={bottomRef} $isInViewBottom={isInViewBottom}>
+          {t('userCount_1')}
+          <S.UserCountWrapper>
+            <S.UserCount>{userCount.toLocaleString()}</S.UserCount>
+            {t('userCount_2')}
+          </S.UserCountWrapper>
+          {t('userCount_3')}
+        </S.UserCountMessage>
+
+        <S.StartButton
+          $isRunningStartTransition={isRunningStartTransition}
+          onClick={handleClickStart}
+        >
+          {t('startButton')}
+        </S.StartButton>
+        <S.StartTransition
+          $isRunningStartTransition={isRunningStartTransition}
+          onTransitionEnd={onStartTransitionEnd}
+        />
+
+        <S.MiniButtonWrapper $isInViewBottom={isInViewBottom}>
+          <S.MiniButton onClick={handleShare}>
+            <FontAwesomeIcon icon={faShareNodes} />
+            {t('shareButton')}
+          </S.MiniButton>
+          <S.MiniButton onClick={handleViewAllType}>
+            <FontAwesomeIcon icon={faSwatchbook} />
+            {t('viewAllTypesButton')}
+          </S.MiniButton>
+        </S.MiniButtonWrapper>
+      </S.BottomPage>
+    </S.PageWrapper>
   );
 }
 
@@ -121,5 +179,3 @@ export const getServerSideProps: GetServerSideProps = async ({ locale }) => {
     },
   };
 };
-
-export default LandingPage;
